Moved ParameterBasedFilterInfoSelectionDataSource to separate file in correct project group
